requirements.txt
streamlit
python-dotenv
langchain-openai
langchain-community
pypdf
faiss-cpu
tiktoken